Убрал генерацию комментариев в отдельный модуль
